+++ s/signals/effect.ts
@@ -2,11 +2,19 @@
 import {debounce} from "@e280/stz"
 import {tracker} from "../tracker/tracker.js"
 
-export function effect(collector: () => void, responder: () => void = collector) {
+export function effect(
+		collector: () => void,
+		responder: () => void = collector,
+	) {
+
 	return collectorEffect(collector, responder).dispose
 }
 
-export function collectorEffect<C = void>(collector: () => C, responder: () => void = collector) {
+export function collectorEffect<C = void>(
+		collector: () => C,
+		responder: () => void = collector,
+	) {
+
 	const {seen, result} = tracker.observe(collector)
 	const fn = debounce(0, responder)