+++ s/signals/index.ts
@@ -1,7 +1,11 @@
 
-export * from "./parts/derive.js"
-export * from "./parts/lazy.js"
+export * from "./parts/reactive.js"
+export * from "./parts/readable.js"
+
+export * from "./derive.js"
 export * from "./effect.js"
+export * from "./fns.js"
+export * from "./lazy.js"
 export * from "./signal.js"
 export * from "./types.js"