+++ s/signals/utils/default-compare.ts
@@ -1,0 +1,5 @@
+
+export function defaultCompare(a: any, b: any) {
+	return a === b
+}
+